WIP: Fix screenshots for Snapshots and Overview
TODO: Overview needs a different scaling
