+++ src/store/model.ts
@@ -208,12 +208,13 @@ export interface IDownloadParameters {
   encoding: string;
   contentDisposition: string;
   contentDispositionFilename: string;
-  left: number;
-  top: number;
-  right: number;
-  bottom: number;
-  width: number;
-  height: number;
+  left?: number;
+  top?: number;
+  right?: number;
+  bottom?: number;
+  width?: number;
+  height?: number;
+  magnification?: number;
   jpeqQuality?: number;
   style?: string;
   tiffCompression?: string;
